Examples, minor tweaks
- added a basic web example
- added comments to the save-as-png example
- changed the default colors to 'random' and 'match_x'
- added tests for Pattern
