+++ examples/basic-web-example.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Trianglify basic example</title>
+    <style>
+    html, body {
+      height: 100%;
+      width: 100%;
+      margin: 0 0;
+      padding: 0 0;
+      text-align: center;
+      cursor: pointer;
+      display: flex;
+      flex-align: center;
+      font-size: 0;
+    }
+
+    canvas {
+      width: 100%;
+      height: 100%;
+    }
+    </style>
+  </head>
+  <body>
+  <script src="../dist/trianglify.min.js"></script>
+  <script>
+    var pattern = Trianglify({
+      height: window.innerHeight,
+      width: window.innerWidth,
+      cell_size: 30+Math.random()*100});
+    document.body.appendChild(pattern.canvas());
+  </script>
+  </body>
+</html>
