+++ lib/trianglify.js
@@ -15,7 +15,7 @@ var Pattern = require('./pattern');
 var defaults = {
   cell_size: 75,                    // Size of the cells used to generate a randomized grid
   variance: 0.75,                   // how much to randomize the grid
-  x_colors: colorbrewer.RdYlBu,     // X color stops
+  x_colors: 'random',               // X color stops
   y_colors: 'match_x',              // Y color stops
   palette: colorbrewer,             // Palette to use for 'random' color option
   color_space: 'lab',               // Color space used for gradient construction & interpolation
@@ -160,7 +160,7 @@ function Trianglify(opts) {
   function _merge_opts(defaults, options) {
     var out = {};
 
-    // shallow-copy defaults
+    // shallow-copy defaults so we don't mutate the input objects (bad)
     for (var key in defaults) {
       out[key] = defaults[key];
     }
